feat: mostly working dnd from dimension panel
