+++ packages/app/src/components/App.css
@@ -78,14 +78,13 @@ body {
     overflow: hidden;
 }
 
-/* .dragging {
-    background-color: yellow;
-} */
-
-.dimItemCopy ~ li {
+/*
+This keeps the clone from causing a ripple effect in the Dimension list
+*/
+.dimension-item-clone ~ li {
     transform: none !important;
 }
 
-.dimItemCopy ~ [data-rbd-placeholder-context-id] {
+/* .dimension-item-clone ~ [data-rbd-placeholder-context-id] {
     display: none !important;
-}
+} */
